Update NuGet package versions and target framework
This commit updates the versions of several NuGet packages in the PCAxis.Menu.UnitTest.csproj and PCAxis.Menu.csproj files. The updates include:
- Microsoft.NET.Test.Sdk: updated from 17.6.2 to 17.9.0
- MSTest.TestAdapter and MSTest.TestFramework: updated from 3.0.4 to 3.3.1
- coverlet.collector: updated from 6.0.0 to 6.0.2

In addition, the target framework of the PCAxis.Menu.UnitTest.csproj file has been upgraded from .NET Core 3.0 to .NET 8.0. These changes aim to improve the functionality and performance of the software.

References to the code changes can be found in the respective .csproj files.